Expect/actual ScatterMap
